SizeAdviserKit/scripts/requirements.txt
GitPython
PyGithub
semver